Remove analytics and restructure
