+++ package.json
@@ -18,15 +18,11 @@
     "test": "jest",
     "lint": "eslint . --ext .ts --ext .tsx --ext .js",
     "type-check": "tsc --noEmit",
-    "build": "rm -rf build && yarn build:analytics && yarn build:render-main",
-    "build:analytics": "ncc build services/analytics/index.js -o build/analytics",
-    "build:render-main": "ncc build services/render-mail/index.js -o build/render-mail",
+    "build": "ncc build src/render-mail/index.js -o build/render-mail",
     "deploy:staging": "serverless deploy -s staging",
     "deploy:production": "serverless deploy -s production",
     "deploy:render-mail:staging": "serverless deploy function -f render-mail -s staging",
-    "deploy:render-mail:production": "serverless deploy function -f render-mail -s production",
     "deploy:analytics:staging": "serverless deploy function -f analytics -s staging",
-    "deploy:analytics:production": "serverless deploy function -f analytics -s production",
     "docker:run": "docker run --env-file=.env sst-services /bin/bash -c",
     "predocker:run": "docker build -t sst-services .",
     "prepare": "husky install"
