--- services/analytics/__mocks__/analytics-report.js
@@ -1,3 +1,0 @@
-const analyticsReport = jest.fn();
-
-module.exports = { analyticsReport };
