--- services/analytics/__tests__/email.js
@@ -1,8 +1,0 @@
-import analyticsContext from '../../../test/mock-data/analytics-context.json';
-import { construct } from '../email';
-
-describe('Module: email.construct', () => {
-  it('should construct an email based on the given context', async () => {
-    expect(await construct(analyticsContext)).toMatchSnapshot();
-  });
-});
