+++ src/render-mail/index.js
@@ -1,11 +1,11 @@
 import { BadRequest, HttpError } from 'http-errors';
 
-import { createResponse } from '../../utils/create-response';
-import * as Env from '../../utils/env';
-import { logger, wrapHandler } from '../../utils/logger';
+import { createResponse } from '../utils/create-response';
+import * as Env from '../utils/env';
+import { logger, wrapHandler } from '../utils/logger';
 import * as Mjml from './mjml';
 
-const render = wrapHandler(async (event, context) => {
+const render = wrapHandler(async (event) => {
   logger.info(`Execution started in env: ${Env.env}`);
 
   try {
