--- static/email-templates/analytics.mjml
@@ -1,209 +1,0 @@
-<mjml>
-  <mj-head>
-    <mj-attributes>
-      <mj-text font-family="Helvetica, Ubuntu, sans-serif" />
-    </mj-attributes>
-
-    <mj-style>
-      a { font-family: Helvetica, Ubuntu, sans-serif; color: #0599e4; text-decoration: none; } a:hover {
-      text-decoration: underline; } .thead { font-weight: normal; text-align: left; } .num { color: #9c9e9f; text-align:
-      right; font-family: monospace; }
-    </mj-style>
-  </mj-head>
-
-  <mj-body background-color="white">
-    {{!-- Header --}}
-    <mj-section padding-bottom="0">
-      <mj-column>
-        <mj-image
-          align="left"
-          width="200px"
-          src="https://www.sjofartstidningen.se/assets/2019/01/ST_logo.png"
-        ></mj-image>
-      </mj-column>
-      <mj-column>
-        <mj-text font-size="24px" padding-bottom="12px">
-          Statistik för webb <br />
-          och nyhetsbrev
-        </mj-text>
-        <mj-text font-size="16px" padding-top="0" color="#9c9e9f">vecka {{ date.week }} – {{ date.year }}</mj-text>
-      </mj-column>
-    </mj-section>
-
-    <mj-section padding-top="0px">
-      <mj-column>
-        <mj-divider border-width="1px"></mj-divider>
-      </mj-column>
-    </mj-section>
-
-    <mj-section padding-top="0px">
-      <mj-column>
-        <mj-text font-size="16px" line-height="24px" color="#9c9e9f">
-          Här är statistik för Sjöfartstidningens hemsida samt nyhetsbrev under perioden {{ date.start }} – {{ date.end
-          }} {{ date.year }}.
-        </mj-text>
-      </mj-column>
-    </mj-section>
-    {{!-- /Header --}} {{!-- Google Analytics --}}
-    <mj-section padding-bottom="10px">
-      <mj-column>
-        <mj-text padding-bottom="0" font-size="24px">Räckvidd webb</mj-text>
-      </mj-column>
-    </mj-section>
-
-    {{#if google}}{{else}}
-    <mj-section>
-      <mj-column>
-        <mj-text font-size="16px" line-height="24px" color="#9c9e9f">
-          Tyvärr gick det inte att hämta data från Google Analytics den här gången
-        </mj-text>
-      </mj-column>
-    </mj-section>
-    {{/if}} {{#with google}}
-    <mj-section padding-top="0px">
-      {{#with this.week}}
-      <mj-column>
-        <mj-text font-size="16px" padding-bottom="0">Senaste veckan:</mj-text>
-        <mj-text padding-bottom="0" font-size="12px" color="#9c9e9f">Jämfört med föregående vecka</mj-text>
-
-        <mj-table>
-          <tr>
-            <th class="thead">Besökare</th>
-            <td class="num">{{formatNumber sessions.total}}</td>
-            <td class="num" style="color: {{#gt sessions.diff 0}}green{{else}}red{{/gt}};">
-              {{percent sessions.diff}}
-            </td>
-          </tr>
-          <tr>
-            <th class="thead">Unika besökare</th>
-            <td class="num">{{formatNumber users.total}}</td>
-            <td class="num" style="color: {{#gt users.diff 0}}green{{else}}red{{/gt}};">{{percent users.diff}}</td>
-          </tr>
-          <tr>
-            <th class="thead">Sidvisningar</th>
-            <td class="num">{{formatNumber pageviews.total}}</td>
-            <td class="num" style="color: {{#gt pageviews.diff 0}}green{{else}}red{{/gt}};">
-              {{percent pageviews.diff}}
-            </td>
-          </tr>
-        </mj-table>
-      </mj-column>
-      {{/with}} {{#with this.month}}
-      <mj-column>
-        <mj-text font-size="16px" padding-bottom="0">Senaste månaden:</mj-text>
-        <mj-text padding-bottom="0" font-size="12px" color="#9c9e9f">Jämfört med samma månad förra året</mj-text>
-
-        <mj-table>
-          <tr>
-            <th class="thead">Besökare</th>
-            <td class="num">{{formatNumber sessions.total}}</td>
-            <td class="num" style="color: {{#gt sessions.diff 0}}green{{else}}red{{/gt}};">
-              {{percent sessions.diff}}
-            </td>
-          </tr>
-          <tr>
-            <th class="thead">Unika besökare</th>
-            <td class="num">{{formatNumber users.total}}</td>
-            <td class="num" style="color: {{#gt users.diff 0}}green{{else}}red{{/gt}};">{{percent users.diff}}</td>
-          </tr>
-          <tr>
-            <th class="thead">Sidvisningar</th>
-            <td class="num">{{formatNumber pageviews.total}}</td>
-            <td class="num" style="color: {{#gt pageviews.diff 0}}green{{else}}red{{/gt}};">
-              {{percent pageviews.diff}}
-            </td>
-          </tr>
-        </mj-table>
-      </mj-column>
-      {{/with}}
-    </mj-section>
-
-    <mj-section>
-      {{#with this.articles}}
-      <mj-column>
-        <mj-text font-size="16px" padding-bottom="0">Mest lästa artiklarna:</mj-text>
-        <mj-table>
-          {{#each this}}
-          <tr>
-            <th class="thead"><a href="{{ url }}">{{capString 160 title}}</a></th>
-            <td class="num">{{formatNumber views}}</td>
-          </tr>
-          {{/each}}
-        </mj-table>
-      </mj-column>
-      {{/with}} {{#with this.jobs}}
-      <mj-column>
-        <mj-text font-size="16px" padding-bottom="0">Mest lästa jobben:</mj-text>
-        <mj-table>
-          {{#each this}}
-          <tr>
-            <th class="thead"><a href="{{ url }}">{{capString 160 title}}</a></th>
-            <td class="num">{{formatNumber views}}</td>
-          </tr>
-          {{/each}}
-        </mj-table>
-      </mj-column>
-      {{/with}}
-    </mj-section>
-    {{/with}} {{!-- /Google Analytics --}}
-
-    <mj-section padding-top="0px">
-      <mj-column>
-        <mj-divider border-width="1px"></mj-divider>
-      </mj-column>
-    </mj-section>
-
-    {{!-- Mailchimp --}}
-    <mj-section padding-bottom="10px">
-      <mj-column>
-        <mj-text padding-bottom="0" font-size="24px">Räckvidd nyhetsbrev</mj-text>
-      </mj-column>
-    </mj-section>
-
-    {{#if mailchimp}}{{else}}
-    <mj-section>
-      <mj-column>
-        <mj-text font-size="16px" line-height="24px" color="#9c9e9f">
-          Tyvärr gick det inte att hämta data från Mailchimp den här gången
-        </mj-text>
-      </mj-column>
-    </mj-section>
-    {{/if}} {{#with mailchimp}}
-    <mj-section padding-top="0px">
-      {{#each this.campaigns}}
-      <mj-column padding-bottom="20px">
-        <mj-text font-size="16px" line-height="19.2px" padding-bottom="0">{{title}}</mj-text>
-        <mj-text font-size="12px" padding-bottom="0" color="#9c9e9f">{{subject}}</mj-text>
-
-        {{#with stats}}
-        <mj-table>
-          <tr>
-            <th class="thead">Mottagare</th>
-            <td class="num">{{formatNumber recipients}}</td>
-          </tr>
-          <tr>
-            <th class="thead">Andel som öppnat</th>
-            <td class="num">{{percent openRate}}</td>
-          </tr>
-          <tr>
-            <th class="thead">Andel som klickat</th>
-            <td class="num">{{percent clickRate}}</td>
-          </tr>
-        </mj-table>
-        {{/with}} {{#with articles}}
-        <mj-text font-size="16px" padding-bottom="0">Mest klickade artiklarna:</mj-text>
-        <mj-table>
-          {{#each this}}
-          <tr>
-            <th class="thead"><a href="{{ url }}">{{capString 160 (urlPath url)}}</a></th>
-            <td class="num">{{formatNumber clicks}}</td>
-          </tr>
-          {{/each}}
-        </mj-table>
-        {{/with}}
-      </mj-column>
-      {{/each}}
-    </mj-section>
-    {{/with}} {{!-- /Mailchimp --}}
-  </mj-body>
-</mjml>
