--- test/mock-data/analytics-context.json
@@ -1,127 +1,0 @@
-{
-  "date": {
-    "week": "1",
-    "year": "2019",
-    "start": "31 december",
-    "end": "6 januari"
-  },
-  "google": {
-    "week": {
-      "users": {
-        "total": 6792,
-        "diff": 0.10528885272579336
-      },
-      "sessions": {
-        "total": 11562,
-        "diff": 0.23591662212720466
-      },
-      "pageviews": {
-        "total": 22942,
-        "diff": 0.24333405592889656
-      }
-    },
-    "month": {
-      "users": {
-        "total": 25759,
-        "diff": -0.03737060428267125
-      },
-      "sessions": {
-        "total": 62325,
-        "diff": -0.09624140831182393
-      },
-      "pageviews": {
-        "total": 128667,
-        "diff": -0.0517437061494016
-      }
-    },
-    "articles": [
-      {
-        "title": "Rekordhög sjö och inställda turer",
-        "url": "https://www.sjofartstidningen.se/rekordhog-sjo-och-installda-turer/",
-        "views": 1368
-      },
-      {
-        "title": "KBV ser över bemanningen",
-        "url": "https://www.sjofartstidningen.se/kbv-ser-bemanningen/",
-        "views": 1044
-      },
-      {
-        "title": "Tyskt klartecken för tunnelbygge",
-        "url": "https://www.sjofartstidningen.se/tyskt-klartecken-tunnelbygge/",
-        "views": 772
-      },
-      {
-        "title": "Ett svinaktigt 2019 stundar",
-        "url": "https://www.sjofartstidningen.se/bloggar/utkiks-bloggen/ett-svinaktigt-2019-stundar/",
-        "views": 513
-      },
-      {
-        "title": "Föreslår utökat isbrytarsamarbete",
-        "url": "https://www.sjofartstidningen.se/sjofartsverket-foreslar-utokat-isbrytarsamarbete/",
-        "views": 505
-      }
-    ],
-    "jobs": [
-      {
-        "title": "Fartygsingenjör",
-        "url": "https://www.sjofartstidningen.se/jobb/fartygsingenjor-3/",
-        "views": 104
-      },
-      {
-        "title": "Befraktare samt två fartygsoperatörer",
-        "url": "https://www.sjofartstidningen.se/jobb/befraktare-samt-tva-fartygsoperatorer/",
-        "views": 101
-      },
-      {
-        "title": "Säljare/Affärsutvecklare",
-        "url": "https://www.sjofartstidningen.se/jobb/saljareaffarsutvecklare-2/",
-        "views": 85
-      },
-      {
-        "title": "Teknisk chef",
-        "url": "https://www.sjofartstidningen.se/jobb/teknisk-chef-8/",
-        "views": 85
-      },
-      {
-        "title": "Överstyrman till ”Astrea”",
-        "url": "https://www.sjofartstidningen.se/jobb/overstyrman-till-astrea/",
-        "views": 69
-      }
-    ]
-  },
-  "mailchimp": {
-    "campaigns": [
-      {
-        "title": "Sjöfartstidningens nyhetsbrev 190103",
-        "subject": "Föreslår utökat isbrytarsamarbete",
-        "stats": {
-          "recipients": 8626,
-          "openRate": 0.35323498664189,
-          "clickRate": 0.13485886862586
-        },
-        "articles": [
-          {
-            "url": "https://www.sjofartstidningen.se/bloggar/utkiks-bloggen/ett-svinaktigt-2019-stundar/?utm_source=Sj%C3%B6fartstidningens+nyhetsbrev&utm_campaign=9a9032bcce-EMAIL_CAMPAIGN_2019_01_03_03_08&utm_medium=email&utm_term=0_eaa9b84478-9a9032bcce-[LIST_EMAIL_ID]",
-            "clicks": 385
-          },
-          {
-            "url": "https://www.sjofartstidningen.se/kbv-ser-bemanningen/?utm_source=Sj%C3%B6fartstidningens+nyhetsbrev&utm_campaign=9a9032bcce-EMAIL_CAMPAIGN_2019_01_03_03_08&utm_medium=email&utm_term=0_eaa9b84478-9a9032bcce-[LIST_EMAIL_ID]",
-            "clicks": 337
-          },
-          {
-            "url": "https://www.sjofartstidningen.se/rekordhog-sjo-och-installda-turer/?utm_source=Sj%C3%B6fartstidningens+nyhetsbrev&utm_campaign=9a9032bcce-EMAIL_CAMPAIGN_2019_01_03_03_08&utm_medium=email&utm_term=0_eaa9b84478-9a9032bcce-[LIST_EMAIL_ID]",
-            "clicks": 302
-          },
-          {
-            "url": "https://www.sjofartstidningen.se/sjofartsverket-foreslar-utokat-isbrytarsamarbete/?utm_source=Sj%C3%B6fartstidningens+nyhetsbrev&utm_campaign=9a9032bcce-EMAIL_CAMPAIGN_2019_01_03_03_08&utm_medium=email&utm_term=0_eaa9b84478-9a9032bcce-[LIST_EMAIL_ID]",
-            "clicks": 266
-          },
-          {
-            "url": "https://www.sjofartstidningen.se/tyskt-klartecken-tunnelbygge/?utm_source=Sj%C3%B6fartstidningens+nyhetsbrev&utm_campaign=9a9032bcce-EMAIL_CAMPAIGN_2019_01_03_03_08&utm_medium=email&utm_term=0_eaa9b84478-9a9032bcce-[LIST_EMAIL_ID]",
-            "clicks": 262
-          }
-        ]
-      }
-    ]
-  }
-}
