--- test/mock-data/google-views-report.json
@@ -1,84 +1,0 @@
-{
-  "reports": [
-    {
-      "columnHeader": {
-        "dimensions": ["ga:pagePath", "ga:pageTitle"],
-        "metricHeader": {
-          "metricHeaderEntries": [
-            {
-              "name": "ga:pageviews",
-              "type": "INTEGER"
-            }
-          ]
-        }
-      },
-      "data": {
-        "rows": [
-          {
-            "dimensions": ["/kbv-ser-bemanningen/", "Sjöfartstidningen | KBV ser över bemanningen |"],
-            "metrics": [
-              {
-                "values": ["1293"]
-              }
-            ]
-          },
-          {
-            "dimensions": [
-              "/rekordhog-sjo-och-installda-turer/",
-              "Sjöfartstidningen | Rekordhög sjö och inställda turer |"
-            ],
-            "metrics": [
-              {
-                "values": ["1136"]
-              }
-            ]
-          },
-          {
-            "dimensions": ["/ater-aktuell-antarktis/", "Sjöfartstidningen | Åter aktuell för Antarktis |"],
-            "metrics": [
-              {
-                "values": ["943"]
-              }
-            ]
-          },
-          {
-            "dimensions": ["/tyskt-klartecken-tunnelbygge/", "Sjöfartstidningen | Tyskt klartecken för tunnelbygge |"],
-            "metrics": [
-              {
-                "values": ["761"]
-              }
-            ]
-          },
-          {
-            "dimensions": [
-              "/skrotningsnivaerna-2018-hogre-pa-30-ar/",
-              "Sjöfartstidningen | Skrotningsnivåerna 2018 högre än på 30 år |"
-            ],
-            "metrics": [
-              {
-                "values": ["642"]
-              }
-            ]
-          }
-        ],
-        "totals": [
-          {
-            "values": ["17904"]
-          }
-        ],
-        "rowCount": 1699,
-        "minimums": [
-          {
-            "values": ["1"]
-          }
-        ],
-        "maximums": [
-          {
-            "values": ["1293"]
-          }
-        ]
-      },
-      "nextPageToken": "5"
-    }
-  ]
-}
