--- test/mock-data/google-visits-report.json
@@ -1,63 +1,0 @@
-{
-  "reports": [
-    {
-      "columnHeader": {
-        "metricHeader": {
-          "metricHeaderEntries": [
-            {
-              "name": "ga:users",
-              "type": "INTEGER"
-            },
-            {
-              "name": "ga:sessions",
-              "type": "INTEGER"
-            },
-            {
-              "name": "ga:pageviews",
-              "type": "INTEGER"
-            }
-          ]
-        }
-      },
-      "data": {
-        "rows": [
-          {
-            "metrics": [
-              {
-                "values": ["8787", "16601", "32792"]
-              },
-              {
-                "values": ["8148", "13787", "27675"]
-              }
-            ]
-          }
-        ],
-        "totals": [
-          {
-            "values": ["8787", "16601", "32792"]
-          },
-          {
-            "values": ["8148", "13787", "27675"]
-          }
-        ],
-        "rowCount": 1,
-        "minimums": [
-          {
-            "values": ["8787", "16601", "32792"]
-          },
-          {
-            "values": ["8148", "13787", "27675"]
-          }
-        ],
-        "maximums": [
-          {
-            "values": ["8787", "16601", "32792"]
-          },
-          {
-            "values": ["8148", "13787", "27675"]
-          }
-        ]
-      }
-    }
-  ]
-}
