--- test/mock-data/mailchimp-campaigns.json
@@ -1,152 +1,0 @@
-{
-  "campaigns": [
-    {
-      "id": "fake_id",
-      "web_id": 111111,
-      "type": "regular",
-      "create_time": "2019-01-03T15:08:50+00:00",
-      "archive_url": "http://eepurl.com/gc03ZH",
-      "long_archive_url": "https://mailchi.mp/sjofartstidningen/freslr-utkat-isbrytarsamarbete",
-      "status": "sent",
-      "emails_sent": 8626,
-      "send_time": "2019-01-03T15:11:19+00:00",
-      "content_type": "html",
-      "needs_block_refresh": false,
-      "resendable": true,
-      "recipients": {
-        "list_id": "fake_id",
-        "list_is_active": true,
-        "list_name": "Sjöfartstidningens nyhetsbrev",
-        "segment_text": "",
-        "recipient_count": 8626
-      },
-      "settings": {
-        "subject_line": "Föreslår utökat isbrytarsamarbete",
-        "title": "Sjöfartstidningens nyhetsbrev 190103",
-        "from_name": "Sjöfartstidningen",
-        "reply_to": "nyhetsbrevet@sjofartstidningen.se",
-        "use_conversation": false,
-        "to_name": "",
-        "folder_id": "",
-        "authenticate": true,
-        "auto_footer": false,
-        "inline_css": false,
-        "auto_tweet": false,
-        "fb_comments": true,
-        "timewarp": false,
-        "template_id": 0,
-        "drag_and_drop": false
-      },
-      "tracking": {
-        "opens": true,
-        "html_clicks": true,
-        "text_clicks": false,
-        "goal_tracking": false,
-        "ecomm360": false,
-        "google_analytics": "EMAIL_CAMPAIGN_2019_01_03_03_08",
-        "clicktale": ""
-      },
-      "report_summary": {
-        "opens": 7070,
-        "unique_opens": 3031,
-        "open_rate": 0.35207341154606,
-        "clicks": 2844,
-        "subscriber_clicks": 1157,
-        "click_rate": 0.13439423858752,
-        "ecommerce": { "total_orders": 0, "total_spent": 0, "total_revenue": 0 }
-      },
-      "delivery_status": { "enabled": false },
-      "_links": [
-        {
-          "rel": "parent",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns",
-          "method": "GET",
-          "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/CollectionResponse.json",
-          "schema": "https://us5.api.mailchimp.com/schema/3.0/CollectionLinks/Campaigns.json"
-        },
-        {
-          "rel": "self",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id",
-          "method": "GET",
-          "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/Response.json"
-        },
-        {
-          "rel": "delete",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id",
-          "method": "DELETE"
-        },
-        {
-          "rel": "send",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/actions/send",
-          "method": "POST"
-        },
-        {
-          "rel": "cancel_send",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/actions/cancel-send",
-          "method": "POST"
-        },
-        {
-          "rel": "feedback",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/feedback",
-          "method": "GET",
-          "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/Feedback/CollectionResponse.json"
-        },
-        {
-          "rel": "content",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/content",
-          "method": "GET",
-          "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/Content/Response.json"
-        },
-        {
-          "rel": "send_checklist",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/send-checklist",
-          "method": "GET",
-          "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/Checklist/Response.json"
-        },
-        {
-          "rel": "pause",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/actions/pause",
-          "method": "POST"
-        },
-        {
-          "rel": "resume",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/actions/resume",
-          "method": "POST"
-        },
-        {
-          "rel": "replicate",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/actions/replicate",
-          "method": "POST"
-        },
-        {
-          "rel": "create_resend",
-          "href": "https://us5.api.mailchimp.com/3.0/campaigns/fake_id/actions/create-resend",
-          "method": "POST"
-        }
-      ]
-    }
-  ],
-  "total_items": 1,
-  "_links": [
-    {
-      "rel": "parent",
-      "href": "https://us5.api.mailchimp.com/3.0/",
-      "method": "GET",
-      "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Root/Response.json"
-    },
-    {
-      "rel": "self",
-      "href": "https://us5.api.mailchimp.com/3.0/campaigns",
-      "method": "GET",
-      "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/CollectionResponse.json",
-      "schema": "https://us5.api.mailchimp.com/schema/3.0/CollectionLinks/Campaigns.json"
-    },
-    {
-      "rel": "create",
-      "href": "https://us5.api.mailchimp.com/3.0/campaigns",
-      "method": "POST",
-      "targetSchema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/Response.json",
-      "schema": "https://us5.api.mailchimp.com/schema/3.0/Definitions/Campaigns/POST.json"
-    }
-  ]
-}
