+++ tsconfig.json
@@ -15,5 +15,5 @@
     "noEmit": true,
     "jsx": "react"
   },
-  "include": ["services", "test", "misc"]
+  "include": ["src"]
 }
