--- utils/width-map.json
@@ -1,99 +1,0 @@
-{
-  "helvetica": {
-    "0": [56, 56, 56, 56],
-    "1": [56, 56, 56, 56],
-    "2": [56, 56, 56, 56],
-    "3": [56, 56, 56, 56],
-    "4": [56, 56, 56, 56],
-    "5": [56, 56, 56, 56],
-    "6": [56, 56, 56, 56],
-    "7": [56, 56, 56, 56],
-    "8": [56, 56, 56, 56],
-    "9": [56, 56, 56, 56],
-    " ": [28, 28, 28, 28],
-    "!": [28, 33, 28, 33],
-    "\"": [35, 47, 35, 47],
-    "#": [56, 56, 56, 56],
-    "$": [56, 56, 56, 56],
-    "%": [89, 89, 89, 89],
-    "&": [67, 72, 67, 72],
-    "'": [19, 24, 19, 24],
-    "(": [33, 33, 33, 33],
-    ")": [33, 33, 33, 33],
-    "*": [39, 39, 39, 39],
-    "+": [58, 58, 58, 58],
-    ",": [28, 28, 28, 28],
-    "-": [33, 33, 33, 33],
-    ".": [28, 28, 28, 28],
-    "/": [28, 28, 28, 28],
-    ":": [28, 33, 28, 33],
-    ";": [28, 33, 28, 33],
-    "<": [58, 58, 58, 58],
-    "=": [58, 58, 58, 58],
-    ">": [58, 58, 58, 58],
-    "?": [56, 61, 56, 61],
-    "@": [102, 98, 102, 98],
-    "A": [67, 72, 67, 72],
-    "B": [67, 72, 67, 72],
-    "C": [72, 72, 72, 72],
-    "D": [72, 72, 72, 72],
-    "E": [67, 67, 67, 67],
-    "F": [61, 61, 61, 61],
-    "G": [78, 78, 78, 78],
-    "H": [72, 72, 72, 72],
-    "I": [28, 28, 28, 28],
-    "J": [50, 56, 50, 56],
-    "K": [67, 72, 67, 72],
-    "L": [56, 61, 56, 61],
-    "M": [83, 83, 83, 83],
-    "N": [72, 72, 72, 72],
-    "O": [78, 78, 78, 78],
-    "P": [67, 67, 67, 67],
-    "Q": [78, 78, 78, 78],
-    "R": [72, 72, 72, 72],
-    "S": [67, 67, 67, 67],
-    "T": [61, 61, 61, 61],
-    "U": [72, 72, 72, 72],
-    "V": [67, 67, 67, 67],
-    "W": [94, 94, 94, 94],
-    "X": [67, 67, 67, 67],
-    "Y": [67, 67, 67, 67],
-    "Z": [61, 61, 61, 61],
-    "[": [28, 33, 28, 33],
-    "\\": [28, 28, 28, 28],
-    "]": [28, 33, 28, 33],
-    "^": [47, 58, 47, 58],
-    "_": [56, 56, 56, 56],
-    "`": [33, 33, 33, 33],
-    "a": [56, 56, 56, 56],
-    "b": [56, 61, 56, 61],
-    "c": [50, 56, 50, 56],
-    "d": [56, 61, 56, 61],
-    "e": [56, 56, 56, 56],
-    "f": [28, 33, 28, 33],
-    "g": [56, 61, 56, 61],
-    "h": [56, 61, 56, 61],
-    "i": [22, 28, 22, 28],
-    "j": [22, 28, 22, 28],
-    "k": [50, 56, 50, 56],
-    "l": [22, 28, 22, 28],
-    "m": [83, 89, 83, 89],
-    "n": [56, 61, 56, 61],
-    "o": [56, 61, 56, 61],
-    "p": [56, 61, 56, 61],
-    "q": [56, 61, 56, 61],
-    "r": [33, 39, 33, 39],
-    "s": [50, 56, 50, 56],
-    "t": [28, 33, 28, 33],
-    "u": [56, 61, 56, 61],
-    "v": [50, 56, 50, 56],
-    "w": [72, 78, 72, 78],
-    "x": [50, 56, 50, 56],
-    "y": [50, 56, 50, 56],
-    "z": [50, 50, 50, 50],
-    "{": [33, 39, 33, 39],
-    "|": [26, 28, 26, 28],
-    "}": [33, 39, 33, 39],
-    "~": [58, 58, 58, 58]
-  }
-}
